+++ src/app/locales/index.ts
@@ -1,5 +1,5 @@
 export * from "./types";
 export * from "./pr-body.builder";
-export * from "./reviewer-warnings-pr.util";
+export * from "./locale-guard-labels.util";
 export * from "./pt-br.locale";
 export * from "./ru.locale";